requirements_macos.txt
loguru
planqk-quantum
pennylane==0.36.0
pennylane-qiskit==0.40.0
numpy>=1.21.6,<1.24.0  # Adjusted for TensorFlow 2.12 compatibility
pandas==2.2.2
scipy==1.10.1  # Ensures compatibility with NumPy
matplotlib==3.9.4
seaborn==0.13.2
scikit-learn==1.2.2  # Downgrade if needed for NumPy compatibility
tensorflow-macos==2.12.0
keras==2.12.0
# tensorflow==2.12.0
# keras==2.12.0
h5py==3.9.0  # Adjusted for NumPy compatibility
boto3==1.36.3
botocore==1.36.3
awscli==1.37.3
requests==2.32.3
protobuf==3.20.3  # Ensure compatibility with TensorFlow
threadpoolctl==3.5.0
joblib==1.4.2
cloudpickle==2.2.1
pydantic==2.7.1
ipykernel==6.29.5
ipython==8.31.0
tensorboard==2.12.0  # Must match TensorFlow version
imbalanced-learn==0.11.0  # Adjusted for NumPy compatibility


typing-extensions>4.6.0,<4.9.0  # Delete-later

# typing-extensions>=4.5.0,<4.9.0  # Compatible with TensorFlow 2.12
